create rudimentary 2018 site
I created a basic 2018 site including a new color scheme and basic nav bar to be filled with content as it comes online.
The old site's _pages have been backed up to _pages/2017.
However, the 2017 site is not part of the jekyll build any more since a pre-compiled version is statically stored in assets/2017.